Rename prometheus_snapshot directory to prometheus
