+++ tests/integration/benchmark_runner/common/oc/test_oc_without_operator.py
@@ -6,8 +6,8 @@ from tests.integration.benchmark_runner.test_environment_variables import *
 import tempfile
 import tarfile
 import time
-from benchmark_runner.common.prometheus_snapshot.prometheus_snapshot import PrometheusSnapshot
-from benchmark_runner.common.prometheus_snapshot.prometheus_snapshot_exceptions import PrometheusSnapshotError, PrometheusSnapshotAlreadyStarted, PrometheusSnapshotNotStarted, PrometheusSnapshotAlreadyRetrieved
+from benchmark_runner.common.prometheus.prometheus_snapshot import PrometheusSnapshot
+from benchmark_runner.common.prometheus.prometheus_snapshot_exceptions import PrometheusSnapshotError, PrometheusSnapshotAlreadyStarted, PrometheusSnapshotNotStarted, PrometheusSnapshotAlreadyRetrieved
 
 
 def test_oc_get_ocp_server_version():
@@ -157,7 +157,7 @@ def test_collect_prometheus():
     oc = OC(kubeadmin_password=test_environment_variable['kubeadmin_password'])
     oc.login()
     with tempfile.TemporaryDirectory() as dirname:
-        snapshot = PrometheusSnapshot(oc=oc, log_path=dirname, verbose=True)
+        snapshot = PrometheusSnapshot(oc=oc, artifacts_path=dirname, verbose=True)
         snapshot.prepare_for_snapshot(pre_wait_time=1)
         time.sleep(10)
         tarball = snapshot.retrieve_snapshot(post_wait_time=1)
